+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - refreshed icons: lightbulb, switch, outlet + unknown
 - refreshed icons: heater, cooler, de/humidifier + thermostat
 - put services types into translatable strings
+- update `monaco-editor` from `v0.21.3` to `v0.52.2`
 
 ### Other Changes